API_LEARN/Google_API/0/pyproject.toml
[project]
name = "google-api-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.56.0",
]
